backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-migrate
psycopg2-binary
recipe-scrapers
requests
beautifulsoup4
